+++ src/layers/l4-controller/commands/index.ts
@@ -10,3 +10,4 @@ export * from './grade';
 export * from './notification';
 export * from './syllabus';
 export * from './sync';
+export * from './settings';